[makeotfexe] garbage in subtable overflow message makeotfexe issue 313
Fix bugs affecting  test files for current code.

The test files for the GPOS mark_to_base tests were failing under Linux because of an old makeotfexe bug in defining anchor lists.  The logic defined a pointer to the previous anchor in an existing list of anchors, and then added a new anchor to the list. The list has a pre-allocated set of anchors; when a new anchor exceeds the length of this list, the entire list may be  re-allocated and moved. If the list is moved, this invalidates in pointers into the prior location. Under Linux, this caused the pointer to the previous anchor to be bad, causing a crash.