[makeotfexe] garbage in subtable overflow message makeotfexe issue 313
The original report was because overflows reported for subtables had unreadable feature tags for two cases: the subtable belonged to a stand-alone named lookup, or to an anonymous lookup created by a chaining contextual rule. This is equally a problem for all the other error reports as well.

First, I standardized the *many* different overflow reports to a single function for each of the  GPOS and GSUB tables.

The I standardized the feature name and lookup names to a single string, 'error_id_text', that is storied in the hotCtx structure, and used it in all the error reports, not just the overflow error reports. All error reports now identify the feature they come from. For both the overflow and other errors, some reports reported this, some didn't.

I then added logic so that this ID text uses the parent feature tag for anonymous lookups, and uses the lookup names for named lookups.

An issue is that some error and overflow reports are encountered only in the GPOSWrite and GSUBWrite phases, when only the subtable info is available, and the context where these are filled no longer exists. In order to allow reporting the feature ( and named lookup), I added a field to the standard Subtable structure to hold the 'error_id_text' string. This field is set when the GPOS and GSUB rules are being filled from the feature file, and is then available for use during the write phase.

Unfortunately, I could not test many of the coverage offsets reports. Because of how makeotfexe  packs the subtables, it is not possible force a coverage table overflow without first encountering the check that one of the lookup subtables overflows the offset limit. I did not remove these checks, as they could be useful if we change how makeotfexe orders the subtables.