[makeotfexe] Clarify path resolution of include statements: issue 164
Fix issue that feature files  referenced from a UFO font won't work in makeotfexe.

Note that for each include statement encountered, the current logic will try in order:
- resolve path relative to main feature file
- if main feature file is in a UFO font, resolve path relative to UFO font directory
- resolve path relative to including file.

This does mean that a font project will build even if it mixes all three models in different include files. The problem is that the is no way to know which model is intended until the first relative include path is encountered that will will work with only one of the models, and it even then it would be possible to draw the wrong conclusion. Ultimately, the most reliable approach  would be to provide options so that a user could force one method or another. I think that scenarios where this will be a problem are sufficiently rare so as to not to be worth dealing with.